+++ android-sdk-ui/src/main/java/com/braze/ui/actions/brazeactions/BrazeActionUtils.kt
@@ -3,8 +3,10 @@
 package com.braze.ui.actions.brazeactions
 
 import android.net.Uri
+import com.appboy.models.cards.Card
 import com.braze.models.inappmessage.IInAppMessage
 import com.braze.models.inappmessage.IInAppMessageImmersive
+import com.braze.models.inappmessage.MessageButton
 import com.braze.ui.actions.brazeactions.BrazeActionParser.ActionType
 import com.braze.ui.actions.brazeactions.BrazeActionParser.getBrazeActionVersionAndJson
 import com.braze.ui.actions.brazeactions.BrazeActionParser.isBrazeActionUri
@@ -17,11 +19,24 @@ import org.json.JSONObject
  * contains a push prompt Braze Action.
  */
 internal fun IInAppMessage.containsAnyPushPermissionBrazeActions(): Boolean =
-    this.getAllUris()
-        .filter { it.isBrazeActionUri() }
-        .mapNotNull { it.getBrazeActionVersionAndJson()?.second }
-        .flatMap { getAllBrazeActionStepTypes(it) }
-        .any { it == ActionType.REQUEST_PUSH_PERMISSION }
+    doAnyTypesMatch(ActionType.REQUEST_PUSH_PERMISSION, this.getAllUris())
+
+/**
+ * Determines whether any [Uri] in the message and/or buttons
+ * contains an invalid action.
+ */
+internal fun IInAppMessage.containsInvalidBrazeAction() =
+    doAnyTypesMatch(ActionType.INVALID, this.getAllUris())
+
+/**
+ * Determines whether any [Uri] in the card contains an invalid action.
+ */
+internal fun Card.containsInvalidBrazeAction(): Boolean {
+    if (this.url != null) {
+        return doAnyTypesMatch(ActionType.INVALID, listOf(Uri.parse(this.url)))
+    }
+    return false
+}
 
 /**
  * Retrieves all [Uri]'s from the main
@@ -56,11 +71,21 @@ internal fun getAllBrazeActionStepTypes(json: JSONObject): List<ActionType> {
                 .forEach { allStepTypes.addAll(getAllBrazeActionStepTypes(it)) }
         }
 
-        // Do nothing if the type is invalid
-        ActionType.INVALID -> {}
-
         // This is an individual/uncontained step
         else -> allStepTypes.add(actionType)
     }
     return allStepTypes
 }
+
+/**
+ * Determines if any of the [Uri] in the list contain the given [ActionType]
+ */
+internal fun doAnyTypesMatch(actionType: ActionType, uriList: List<Uri>): Boolean =
+    uriList
+        .filter { it.isBrazeActionUri() }
+        // If the parsed json is null, then return an empty json
+        // object so we can return it as invalid in
+        // the next step.
+        .map { it.getBrazeActionVersionAndJson()?.second ?: JSONObject() }
+        .flatMap { getAllBrazeActionStepTypes(it) }
+        .any { it == actionType }