+++ android-sdk-ui/src/main/java/com/braze/ui/actions/brazeactions/steps/BaseBrazeActionStep.kt
@@ -1,21 +1,19 @@
 package com.braze.ui.actions.brazeactions.steps
 
-import com.braze.events.SimpleValueCallback
 import com.braze.Braze
 import com.braze.BrazeUser
+import com.braze.events.IValueCallback
 import com.braze.support.BrazeLogger.brazelog
 
 sealed class BaseBrazeActionStep : IBrazeActionStep {
     companion object {
         internal fun Braze.runOnUser(block: (user: BrazeUser) -> Unit) {
-            this.getCurrentUser(object : SimpleValueCallback<BrazeUser>() {
-                override fun onSuccess(user: BrazeUser) {
-                    super.onSuccess(user)
-                    block(user)
+            this.getCurrentUser(object : IValueCallback<BrazeUser> {
+                override fun onSuccess(value: BrazeUser) {
+                    block(value)
                 }
 
                 override fun onError() {
-                    super.onError()
                     brazelog { "Failed to run on Braze user object" }
                 }
             })