+++ android-sdk-ui/src/main/java/com/braze/ui/contentcards/handlers/DefaultContentCardsUpdateHandler.kt
@@ -4,15 +4,15 @@ import android.os.Parcel
 import android.os.Parcelable
 import com.appboy.models.cards.Card
 import com.braze.events.ContentCardsUpdatedEvent
+import com.braze.ui.actions.brazeactions.containsInvalidBrazeAction
 import java.util.*
 
 class DefaultContentCardsUpdateHandler : IContentCardsUpdateHandler {
     override fun handleCardUpdate(event: ContentCardsUpdatedEvent): List<Card> {
-        val sortedCards: List<Card> = event.allCards
         // Sort by pinned, then by the 'updated' timestamp descending
         // Pinned before non-pinned
-        Collections.sort(sortedCards) { cardA: Card, cardB: Card ->
-            return@sort when {
+        val cardComparator = Comparator { cardA: Card, cardB: Card ->
+            when {
                 // A displays above B since A is pinned and B isn't
                 cardA.isPinned && !cardB.isPinned -> -1
                 // B displays above A since B is pinned and A isn't
@@ -26,7 +26,8 @@ class DefaultContentCardsUpdateHandler : IContentCardsUpdateHandler {
                 else -> 0
             }
         }
-        return sortedCards
+        return event.allCards.filter { card -> !card.containsInvalidBrazeAction() }
+            .sortedWith(cardComparator)
     }
 
     // Parcelable interface method